Functionality of Tic Tac Toe
Admittedly, this isn't the right way to do things, but I was spiking this out
first because I didn't know how Typescript and Express, etc worked together.

But... this works! app/index.ts holds the "router" and app/games.ts holds the
"actions" which farms off to the "model" in app/model/game.ts. There is no
persistence but there is the allowance for multiple games, although the ids
are predictable and reset on server restart (but so do the games themselves so
it's not that big a deal).

Since this is spiking the whole thing, it also includes TS config/formatting
files.

Run the server:
`npm install`
`npm start` (change port by modifying `.env` or, I believe, `PORT=x npm start`
should work too), port 3001 by default.